all commentaries solved + app start methods added

+++ src/main/java/sqltask/courses/CoursesParser.java
@@ -1,5 +1,6 @@
 package sqltask.courses;
 
+import org.apache.ibatis.jdbc.Null;
 import sqltask.helpers.*;
 
 public class CoursesParser implements Parser<Course> {
@@ -7,6 +8,6 @@ public class CoursesParser implements Parser<Course> {
     @Override
     public Course parse(String str) {
         String[] parts = str.split("_");
-        return new Course(null, parts[1], parts[2]);
+        return new Course((Integer) null, parts[1], parts[2]);
     }
 }
